Move Vec2 into it's own file (GLM removal)
