+++ simulant.includes
@@ -130,3 +130,4 @@ deps/bounce/src/bounce/dynamics
 simulant/utils/mesh
 simulant/sound_drivers
 tests/gl2
+simulant/math
